Revert local changes to number of hits
The commands to set up the Dockerised PostgreSQL database in
provider/postgresql.py configure it with an external volume for the
data.  Running them repeatedly as part of a script caused multiple sets
of the OSM data to be added into the tables and changed the numbers for
the number of hits query, which meant that the tests had to be modified
to pass.  This commit reverts the number of hits back to the original
value.