Json serialize numpy types (#913)
* Json serialize numpy types

This commit fixes an issue with the current xarray provider. When a
NetCDF file is loaded into an xarray dataset, attributes of data
variables that are numbers will be loaded as numpy.float64 or
numpy.int64. These attributes are passed directly to metadata objects
that are json serialized, creating the issue.

This commit thereby fixes #895

* Avoid numpy import in util module