fix(OGRProvider_init): make the 'id_field' attribute mandatory in provider config (#1135)
As for the 'layer' attribute, the 'id_field' should be specified in the
configuration file for the OGRProvider. The 'get' method of a OGRProvider
instance (e.g. when requesting a single feature) needs the feature ID when
setting an attribute filter on the layer. If the 'id_field' attribute is
omitted in the configuration file, the error thrown is more explicit.