OGC API Part4 Create/POST: accept incoming feature without id
The spec mentions:

Requirement 4: /req/create-replace-delete/insert-response-rid
If the operation completes successfully, the server SHALL assign
a new, unique identifier within the collection for the newly added resource.

and

Permission 4: /per/create-replace-delete/rid
If the representation of the resource submitted in the request body
contained a resource identifier, the server MAY use this identifier
as the new resource identifier in the collection or the server MAY
ignore the value and assign its own identifier for the resource.

Currently pygeopapi implements the first part of permission 4 (accepting
and using the identifier from the incomping feature), but errors out
when it is missing, which is against requirement 4.