Case-insensitive headers for execute_process (#1256)
* Address case-insensitive for headers in execute_process

Headers fields are general case insensitive. I found my client was sending the prefer field as lower case, resulting in my processes being run synchronously. This commit changes the default value of the dict.get command for 'Prefer' to try looking up 'prefer' if it cannot find the capital version of the field in the headers dictionary.

* Address flake8 issues

Tweaks to style to address flake8 flagged issues.