Use same format handling in collection coverage as elsewhere
This now fixes the issue that the `Accept` header wasn't considered for
format selection.

This adds netcdf as a known mime type in order to be able to test this,
but also for actual clients to use.

Note that if we want to allow other mime types such as `application/x-grib2`,
we also have to add them to `FORMAT_TYPES`.

Fixes #1390