Clear up space before running CI
With all the databases and dependencies more than the 19G is
needed (this is the amount which is currenlty available on runners).

There is a github action which meddles with the disc layout, which is
probably too much for our use case, but it gave me the idea that you can
delete files which are needed to support certain platforms such as
dotnet and android.
https://github.com/easimon/maximize-build-space

With only manually removing those files, around 30G are available which
should be enough for the foreseeable future